Fix scroll bar and header not updating on scrollThis commit fixes the issue where the scroll bar and header were not updating when scrolling. It adds a scroll event listener to update the scrollTop value and passes it to the respective components
